adding qpus from qprov
